feat 新仕様にともなうページの整理
