+++ src/pages/project-list.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ProjectList = () => {
+  return <div>projectList</div>;
+};
+
+export default ProjectList;
